Add support for single IPU keras models.
Summary:
Provide delegate functions for fit/predict/eval in Keras which works for sequential and functional models.
This currently doesn't support gradient accumulation - this will be done as a follow up diff.
TF2.4 Only
Ref T41125

Test Plan: Added new tests, adjusted existing tests

Reviewers: jackh, alfiee, markf, samuelh, #tensorflow, simonl, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: jackh, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T41125

Differential Revision: https://phabricator.sourcevertex.net/D47943

+++ tensorflow/python/ops/nn_impl.py
@@ -1633,7 +1633,7 @@ def fused_batch_norm(
   if mean is None:
     mean = constant_op.constant(dtype=x.dtype, value=[])
   if variance is None:
-    variance = constant_op.constant([])
+    variance = constant_op.constant([], dtype=scale.dtype)
 
   # Set a minimum epsilon to 1.001e-5, which is a requirement by CUDNN to
   # prevent exception (see cudnn.h).
